vim: Trim leading whitespace from clipboard copied text
When copying to the system clipboard in vim, it's often to share code.
Most of the times when I do this I don't want to include all the leading
whitespace if I've copied a block of code that's deeply nested

For example, if I have the code

```
if testing {
  while condition {
    // There's a bug in here, can anyone help?
    code_with_bug()
  }
}
```

If I wanted to yank the two inner most lines into my clipboard to share,
I would end up with this in my clipboard

```
    // There's a bug in here, can anyone help?
    code_with_bug()
```

Really what I want is this in my clipboard

```
// There's a bug in here, can anyone help?
code_with_bug()
```

This addition allows for that by adjusting my system clipboard mapping
to instead place the text into the "c register. An autocmd will then
pickup yanks into that register and trim the white space.